Move parameters tests
But still need to add a _lot_ of tests!
